decouple ComplexNodeTypeResolver, add PHP 7.4 type resolving from assigns
